Products Creation and Overview incl. Database finished

+++ DogginatorLibrary/Models/ProductModel.cs
@@ -17,7 +17,7 @@ namespace DogginatorLibrary.Models
         public int ItemNumber { get; set; }
         public string Shortdescription { get; set; }
         public string Longdescription { get; set; }
-        public float Price { get; set; }
+        public Decimal Price { get; set; }
         public string Create_Date { get; set; }
         public string Edit_Date { get; set; }
         #endregion
